+++ apps/docs/tailwind.config.ts
@@ -1,10 +1,10 @@
 import type { Config } from "tailwindcss";
-import aristobyteui from "@aristobyte-ui/theme";
+import { aristobyteui } from "@aristobyte-ui/theme";
 
 const config: Config = {
   content: ["./app/**/*.{ts,tsx}", "./components/**/*.{ts,tsx}"],
   theme: { extend: { ...aristobyteui.extend() } },
-  plugins: [...aristobyteui.plugin()],
+  plugins: [...aristobyteui.plugins()],
 };
 
 export default config;