+++ packages/theme/extend/index.ts
@@ -1,25 +1,4 @@
-import { transitionTimingFunction } from "./animations";
-import { colors } from "./colors";
-import { fontFamily, fontSize, fontWeight, lineHeight } from "./fonts";
-import { boxShadow } from "./shadows";
-
-export * from "./animations";
-export * from "./colors";
-export * from "./fonts";
-export * from "./shadows";
-export default () => ({
-  // === Colors === //
-  colors,
-
-  // === Typography === //
-  fontFamily,
-  fontWeight,
-  lineHeight,
-  fontSize,
-
-  // === Shadows === //
-  boxShadow,
-
-  // === Animations === //
-  transitionTimingFunction,
-});
+export { transitionTimingFunction } from "./animations";
+export { colors } from "./colors";
+export { fontFamily, fontSize, fontWeight, lineHeight } from "./fonts";
+export { boxShadow } from "./shadows";