+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @aristobyte-ui/cli
 
+## 1.0.88
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aristobyte-ui/eslint-config@1.0.88
+  - @aristobyte-ui/typescript-config@1.0.88
+
+## 1.0.87
+
+### Patch Changes
+
+- fix: resolve issues with tsup
+
+- Updated dependencies []:
+  - @aristobyte-ui/typescript-config@1.0.87
+  - @aristobyte-ui/eslint-config@1.0.87
+
+## 1.0.85
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aristobyte-ui/eslint-config@1.0.85
+  - @aristobyte-ui/typescript-config@1.0.85
+
 ## 1.0.84
 
 ### Patch Changes